Fix IdentityCenter access rules not being updated
